chore: Update image dimensions and styling in CSS
Adjust the width and height of the campfest.svg image in the CSS file to 250px and 125px respectively. Also, add a margin-left of 10px to the image. Additionally, update the navbar.ejs file to set the height of the navbar to 120px and add border radius to the bottom corners. These changes improve the visual appearance and layout of the website.