chore: Update image property in campground model and views
The image property in the campground model and views has been updated to support multiple images. It is now an array of objects, each containing a URL and filename. This modification allows for easier handling and display of multiple images for each campground. The image is being store in Cloudinary and the link to the image is being store in MongoDB.  Also update the show, new and edit pages so that it can upload multiple images, see multiple images and adding new images when updating campground.