refactor: Update campground image handling and deletion
This commit updates the handling of campground images and introduces the ability to delete images when updating a campground. The image property in the campground model and views has been refactored to support multiple images, stored as an array of objects with URL and filename. Additionally, the deleteImages property has been added to the campgroundSchema to store images that need to be deleted. This modification improves the management and display of multiple images for each campground. The image is stored in Cloudinary and the link to the image is stored in MongoDB. The show, new, and edit pages have also been updated to allow for uploading, viewing, and adding new images when updating a campground. Also, Adding virtual thumbnail to image.url so that easy to make transformation to the image uploads.