feat: Add Mapbox SDK dependency and implement map display on campground page
This commit adds the Mapbox SDK dependency to the project by updating the package.json file. It also implements the display of a map on the campground page using the Mapbox API. The map is centered on the coordinates of the campground and a marker is added to indicate the location. This enhancement improves the user experience by providing a visual representation of the campground's location.