refactor: Update campground creation to use Malaysia campsite data
This commit modifies the `index.js` file to update the campground creation process. Instead of using random descriptors and places, the campgrounds now use data from the `malaysiaCampsite.js` file. The `title` and `location` fields are populated with the corresponding values from the `malaysiaCampsite` array. This change improves the accuracy and relevance of the campground data, providing a better user experience.