refactor: Update image thumbnail height and border radius
The code changes in the `style.css` file update the height and border radius of the `#image_thumbnail` element. The height is increased to 700px, and the border radius is modified to have a rounded top-left corner and a rounded bottom-right corner. These changes improve the visual appearance of the image thumbnail.